inline into test_specifier()/bench_specifier()
